+++ .vscode/c_cpp_properties.json
@@ -1,0 +1,23 @@
+{
+    "version": 4,
+    "configurations": [
+        {
+            "name": "Windows",
+            "includePath": [
+                "${workspaceFolder}/include",  // Dein Include-Ordner
+                "${workspaceFolder}/src",     // Dein src-Ordner (falls relevant)
+                "${workspaceFolder}/**"       // Alle Unterordner durchsuchen
+            ],
+            "defines": [],
+            "compilerPath": "C:\\Program Files\\mingw-w64\\x86_64-8.1.0-posix-seh-rt_v6-rev0\\mingw64\\bin\\gcc.exe",  // Beispiel: MinGW Compiler
+            "intelliSenseMode": "gcc-x64", // IntelliSense für GCC
+            "browse": {
+                "path": [
+                    "${workspaceFolder}/include",
+                    "${workspaceFolder}/src"
+                ],
+                "limitSymbolsToIncludedHeaders": true
+            }
+        }
+    ]
+}
