+++ src/wortsuche.cpp
@@ -1,59 +1,9 @@
-#include <random>
-#include <algorithm>
 #include <iostream>
-#include <mutex>
 #include <thread>
 
 #include "wortsuche.h"
 
 using namespace std;
-
-
-TrieNode::TrieNode() : EndOfWord(false) {
-    for(int i=0; i<NUMBER_OF_LETTERS; ++i){
-        children[i] = nullptr;
-    }
-}
-
-void insert(TrieNode* root, string& key){
-    TrieNode* current = root;
-    string parent_data;
-    
-    for (int i=0; i<key.length(); ++i){
-        char c = key[i];
-        if (current->children[c-'A'] == nullptr){
-            TrieNode* newNode = new TrieNode();
-            current->children[c-'A'] = newNode;
-        }
-        current = current->children[c-'A'];
-        string inter(1,c);
-        current->data = parent_data + c;
-        parent_data = parent_data + c;
-    }
-    current->EndOfWord = true;
-}
-
-void generateWords(int length, const string& currentWord, vector<string>& wordList){
-    if(currentWord.size() == length){
-        wordList.push_back(currentWord);
-        return;
-    }
-
-    for(char c='A'; c<='Z'; c++){
-        generateWords(length, currentWord+c, wordList);
-    }
-}
-
-vector<string> createWordListRecursive(int length){
-    random_device gen;
-    mt19937 g(gen());
-    vector<string> wordList;
-
-    generateWords(length, "", wordList);
-    shuffle(wordList.begin(), wordList.end(), gen);
-
-    return wordList;
-}
 
 void searchHelper(TrieNode* root, vector<TrieNode*>& threadNode){
     vector<TrieNode*> localNodes;
@@ -117,45 +67,6 @@ vector<string> searchFinal(TrieNode* root, string& key){
         for(TrieNode* t : Nodes){
             FoundWords.push_back(t->data);
         }
-    }
-    return FoundWords;
-}
-
-
-
-
-
-void searchHelperWithoutThreads(TrieNode* root, vector<TrieNode*>& Nodes, vector<string>& foundWords){
-    vector<TrieNode*> localNodes;
-    for(char x = 'A'; x<='Z'; x++){
-        if(root->children[x-'A'] != nullptr){
-            localNodes.push_back(root->children[x-'A']);
-        }
-    }
-    Nodes.insert(Nodes.end(), localNodes.begin(), localNodes.end());
-    for(TrieNode* t : localNodes){
-        if(t->EndOfWord) foundWords.push_back(t->data);
-    }
-}
-
-
-vector<string> searchFinalWithoutThreads(TrieNode* root, string& key){
-    TrieNode* current = root;
-    vector<string> FoundWords;
-    vector<TrieNode*> Nodes;
-
-    for(char c : key){
-        if(current->children[c-'A'] == nullptr) return FoundWords;
-        current = current->children[c-'A'];
-    }
-    if(current->EndOfWord) FoundWords.push_back(current->data);
-
-    searchHelperWithoutThreads(current, Nodes, FoundWords);
-    while(!Nodes.empty()){
-        vector<TrieNode*> Intermediate_nodes = Nodes;
-        Nodes.clear();
-
-        for(TrieNode* inter : Intermediate_nodes){searchHelperWithoutThreads(inter, Nodes, FoundWords);}
     }
     return FoundWords;
 }
